refactor(api): revert dataLog.adapter which is not in .port
